ai-tax-management-system/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-cosmos>=4.14.3",
    "azure-servicebus>=7.14.3",
    "fastapi>=0.124.4",
    "loguru>=0.7.3",
    "minio>=7.2.20",
    "openpyxl>=3.1.0",
    "pika>=1.3.2",
    "pypdf>=5.1.0",
    "python-multipart>=0.0.20",
    "semantic-kernel>=1.36.0",
    "uvicorn>=0.38.0",
]
